// src/types/enums.ts
export enum TodoListActions  {
    ADD_TODO = "ADD_TODO_ITEM",
    CHECK_TODO = "CHECK_TODO_ITEM",
    DLETE_TODO = "DELETE_TODO_ITEM",
}


export enum IS_Check {
    TRUE = 1,
    FALSE = 0,
}


export enum FILTER {
    ALL = "ALL",
    COMPLETED = "COMPLETED",
    NOT_COMPLETED = "NOT_COMPLETED"
}